Current state

+++ signals.cpp
@@ -11,13 +11,14 @@
 void signal_handle(int signal) 
 {
     list<job>::iterator it;
-    if (fg_job != NULL)
+    if (fg_job != NULL && fg_job->is_fg == true)
     {
         switch(signal)
         {
-            case (SIGTSTP):
-                cout << "SIGTSTP" << endl;
+            case (SIGTSTP): // CTR-Z
+                fprintf(stdout, "smash > signal SIGTSTP was sent to %d\n", fg_job->pid); // TODO fix message
 	            kill(fg_job->pid,SIGTSTP);
+                fg_job->is_running = false;
                 fg_job->is_fg = false;
                 for( it = jobs->begin() ; it != jobs->end() ; it++)
                     if ((*it).pid == fg_job->pid)
@@ -27,14 +28,13 @@ void signal_handle(int signal)
                     new_job.id = job_cnt++; 
                     jobs->push_back(new_job);
                     fg_job = &(jobs->back());
-                    new_job.id = 0;
-                    cout << "Inserted to the jobs list" << endl;
                 }
 
                 break;
-            case (SIGINT):
-                cout << "SIGINT" << endl;
-	            kill(fg_job->pid,SIGINT);
+            case (SIGINT): // CTR-C
+                    fg_job->is_fg = false;
+                    fprintf(stdout, "smash > signal SIGINT was sent to %d\n", fg_job->pid); // TODO fix message
+	                kill(fg_job->pid,SIGINT);
                 break;
             default:
                 break;
@@ -45,29 +45,32 @@ void signal_handle(int signal)
 void child_handle(int signal)
 {
     pid_t pid = waitpid(-1, NULL, WNOHANG);
-    cout << "SIGCHLD PID IS:" << pid << endl;
-    if (fg_job->pid == pid)
-        fg_job = NULL;
-    if (new_job.pid == pid)
+    if (pid > 0)
     {
-        cout << "job " << pid << "erased" << endl;
-        new_job.id = 0;
-        new_job.cmd = "";
-        new_job.pid = 0;
-        new_job.is_fg = 0;
+        if (fg_job != NULL && fg_job->pid == pid)
+            fg_job = NULL;
+        /*
+        if (new_job.pid == pid)
+        {
+            new_job.id = 0;
+            new_job.cmd = "";
+            new_job.pid = 0;
+            new_job.is_running = 0;
+        }
+        */
     }
     while(pid > 0)
     {       
         list<job>::iterator it;
-        cout << "fg job pid: " << fg_job->pid << endl;
         for( it = jobs->begin() ; it != jobs->end() ; it++)
+        {
             if ((*it).pid == pid)
             {
                 jobs->erase(it);
                 break;
             }
+        }
         pid = waitpid(-1, NULL, WNOHANG);
-        cout << "SIGCHLD PID IS:" << pid << "Is next" << endl;
     }
 }
